Integration tests for delete mappings

See merge request comp23412_2020/eventlite_G04!29

+++ src/main/java/uk/ac/man/cs/eventlite/controllers/EventsController.java
@@ -104,6 +104,11 @@ public class EventsController {
 	}
 	@DeleteMapping("/{eventId}")
 	public String deleteEvent(@PathVariable long eventId, RedirectAttributes redirectAttrs) {
+		Optional<Event> e = eventService.findEventById(eventId);
+		if (e.isEmpty()) {
+			redirectAttrs.addFlashAttribute("error_message", "Event doesn't exist anymore!");
+			return "redirect:/events";
+		}
 		eventService.deleteById(eventId);
 		redirectAttrs.addFlashAttribute("ok_message", "Event deleted.");
 		return "redirect:/events";
